// apps/api/src/modules/auth/auth.schema.ts
import { z } from "zod";

export const registerSchema = z.object({
  email: z.string().email(),

  password: z.string().min(6),

  name: z.string(),
});

export const loginSchema = z.object({
  email: z.string().email(),

  password: z.string(),
});

export type RegisterPayload = z.infer<typeof registerSchema>;

export type LoginPayload = z.infer<typeof loginSchema>;